[jif]: clippy lints

+++ jif/src/read/jif.rs
@@ -59,7 +59,7 @@ impl JifRaw {
         let to_skip =
             header.itrees_size as i64 - (n_itree_nodes * ITreeNode::serialized_size()) as i64;
         let itree_nodes = (0..n_itree_nodes)
-            .map(|idx| ITreeNode::from_reader(r, idx as usize))
+            .map(|idx| ITreeNode::from_reader(r, idx))
             .collect::<Result<Vec<_>, _>>()?;
         r.seek_relative(to_skip)?;
